Some more InteliiJ lead improvements, this time taking advantage of all the most modern Java syntax up to Java 1.8
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20605 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/request/CrimboCafeRequest.java
@@ -102,7 +102,7 @@ public class CrimboCafeRequest
 		for ( int i = 0; i < CrimboCafeRequest.MENU_DATA.length; ++i )
 		{
 			Object [] data = CrimboCafeRequest.MENU_DATA[ i ];
-			if ( itemId == ((Integer)data[ 1 ]).intValue() )
+			if ( itemId == ( Integer ) data[ 1 ] )
 			{
 				return data;
 			}
@@ -117,12 +117,12 @@ public class CrimboCafeRequest
 
 	public final static int dataItemID( Object [] data )
 	{
-		return ((Integer)data[ 1 ]).intValue();
+		return ( Integer ) data[ 1 ];
 	}
 
 	public final static int dataPrice( Object [] data )
 	{
-		return ((Integer)data[ 2 ]).intValue();
+		return ( Integer ) data[ 2 ];
 	}
 
 	public CrimboCafeRequest( final String name )
